+++ rust/operator-binary/src/crd/mod.rs
@@ -1,6 +1,7 @@
 pub mod affinity;
 pub mod authentication;
 pub mod catalog;
+pub mod client_protocol;
 pub mod discovery;
 pub mod fault_tolerant_execution;
 
@@ -61,6 +62,7 @@ pub const LOG_PROPERTIES: &str = "log.properties";
 pub const ACCESS_CONTROL_PROPERTIES: &str = "access-control.properties";
 pub const JVM_SECURITY_PROPERTIES: &str = "security.properties";
 pub const EXCHANGE_MANAGER_PROPERTIES: &str = "exchange-manager.properties";
+pub const SPOOLING_MANAGER_PROPERTIES: &str = "spooling-manager.properties";
 // node.properties
 pub const NODE_ENVIRONMENT: &str = "node.environment";
 // config.properties
@@ -103,6 +105,7 @@ pub const STACKABLE_MOUNT_INTERNAL_TLS_DIR: &str = "/stackable/mount_internal_tl
 pub const STACKABLE_TLS_STORE_PASSWORD: &str = "changeit";
 // secret vars
 pub const ENV_INTERNAL_SECRET: &str = "INTERNAL_SECRET";
+pub const ENV_SPOOLING_SECRET: &str = "SPOOLING_SECRET";
 // TLS
 pub const TLS_DEFAULT_SECRET_CLASS: &str = "tls";
 // Logging
@@ -297,6 +300,10 @@ pub mod versioned {
         #[serde(skip_serializing_if = "Option::is_none")]
         pub fault_tolerant_execution:
             Option<fault_tolerant_execution::FaultTolerantExecutionConfig>,
+
+        /// Client spooling protocol configuration.
+        #[serde(skip_serializing_if = "Option::is_none")]
+        pub client_protocol: Option<client_protocol::ClientProtocolConfig>,
 
         /// Name of the Vector aggregator [discovery ConfigMap](DOCS_BASE_URL_PLACEHOLDER/concepts/service_discovery).
         /// It must contain the key `ADDRESS` with the address of the Vector aggregator.